Refactor to modular Go architecture with clean package separation
- Split monolithic main.go into focused packages (models, processor, ui, utils)
- Create type-safe data structures in models package
- Separate core processing logic into processor package
- Extract UI functionality into dedicated ui package with interactive mode
- Organize utility functions in utils package with file operations
- Update documentation to reflect modular architecture
- Maintain all existing functionality while improving code organization
- Add specific file references and line counts to documentation
