+++ go.mod
@@ -1,3 +1,3 @@
-module palo-pan-parsing
+module github.com/jtbwatson/palo-pan-parsing
 
 go 1.20
