+++ utils/writer.go
@@ -1,0 +1,288 @@
+package utils
+
+import (
+	"fmt"
+	"os"
+	"sort"
+	"strings"
+
+	"github.com/jtbwatson/palo-pan-parsing/models"
+)
+
+// WriteResults writes results to file in a structured YAML-like format
+func WriteResults(outputFile, addressName string, matchingLines []string, itemsDict *models.FormattedResults) error {
+	// Ensure outputs directory exists
+	if err := EnsureOutputsDir(); err != nil {
+		return fmt.Errorf("error creating outputs directory: %w", err)
+	}
+	
+	// Prepend outputs/ to the file path if not already there
+	if !strings.HasPrefix(outputFile, "outputs/") {
+		outputFile = "outputs/" + outputFile
+	}
+	
+	file, err := os.Create(outputFile)
+	if err != nil {
+		return fmt.Errorf("error creating output file: %w", err)
+	}
+	defer file.Close()
+
+	// Enhanced header with metadata
+	fmt.Fprintf(file, "# ═══════════════════════════════════════════════════════════════\n")
+	fmt.Fprintf(file, "# PAN Log Parser Analysis Report v2.0 (Go Edition)\n")
+	fmt.Fprintf(file, "# ═══════════════════════════════════════════════════════════════\n")
+	fmt.Fprintf(file, "# Target Address Object: %s\n", addressName)
+	fmt.Fprintf(file, "# Configuration Lines Found: %d\n", len(matchingLines))
+
+	totalRelationships := len(itemsDict.DeviceGroups) + len(itemsDict.DirectSecurityRules) +
+		len(itemsDict.IndirectSecurityRules) + len(itemsDict.AddressGroups) +
+		len(itemsDict.NATRules) + len(itemsDict.ServiceGroups) + len(itemsDict.RedundantAddresses)
+	fmt.Fprintf(file, "# Total Relationships: %d\n", totalRelationships)
+	fmt.Fprintf(file, "# ═══════════════════════════════════════════════════════════════\n\n")
+
+	// Matching configuration lines section
+	fmt.Fprintf(file, "# MATCHING CONFIGURATION LINES\n")
+	fmt.Fprintf(file, "# Found %d lines containing '%s'\n", len(matchingLines), addressName)
+	fmt.Fprintf(file, "---\n\n")
+
+	if len(matchingLines) > 0 {
+		for i, line := range matchingLines {
+			fmt.Fprintf(file, "  %2d. %s\n", i+1, line)
+		}
+	} else {
+		fmt.Fprintf(file, "  # No matching lines found\n")
+	}
+
+	// Category sections
+
+	// Write each category
+	writeCategory(file, "Device Groups", itemsDict.DeviceGroups)
+	writeSecurityRulesCategory(file, "Direct Security Rules", itemsDict.DirectSecurityRules)
+	writeSecurityRulesCategory(file, "Indirect Security Rules (via Address Groups)", itemsDict.IndirectSecurityRules)
+	writeAddressGroupsCategory(file, itemsDict.AddressGroups)
+	writeCategory(file, "NAT Rules", itemsDict.NATRules)
+	writeCategory(file, "Service Groups", itemsDict.ServiceGroups)
+	writeRedundantAddressesCategory(file, itemsDict.RedundantAddresses)
+
+	// Add footer
+	fmt.Fprintf(file, "# ═══════════════════════════════════════════════════════════════\n")
+	fmt.Fprintf(file, "# Analysis Complete\n")
+	fmt.Fprintf(file, "# Generated by: PAN Log Parser Tool v2.0 (Go Edition)\n")
+	fmt.Fprintf(file, "# Advanced Palo Alto Networks Configuration Analysis\n")
+	fmt.Fprintf(file, "# ═══════════════════════════════════════════════════════════════\n")
+
+	return nil
+}
+
+func writeCategory(file *os.File, category string, items []string) {
+	count := len(items)
+	fmt.Fprintf(file, "\n# %s\n", strings.ToUpper(category))
+	fmt.Fprintf(file, "# Found: %d item", count)
+	if count != 1 {
+		fmt.Fprintf(file, "s")
+	}
+	fmt.Fprintf(file, "\n---\n")
+
+	if count > 0 {
+		for i, item := range items {
+			fmt.Fprintf(file, "  %d. %s\n", i+1, item)
+		}
+	} else {
+		fmt.Fprintf(file, "  None discovered\n")
+	}
+	fmt.Fprintf(file, "\n")
+}
+
+func writeSecurityRulesCategory(file *os.File, category string, items []string) {
+	count := len(items)
+	fmt.Fprintf(file, "\n# %s\n", strings.ToUpper(category))
+	fmt.Fprintf(file, "# Found: %d item", count)
+	if count != 1 {
+		fmt.Fprintf(file, "s")
+	}
+	fmt.Fprintf(file, "\n---\n")
+
+	if count > 0 {
+		// Group rules by device group
+		rulesByDG := make(map[string][]models.RuleContext)
+		for _, item := range items {
+			parts := strings.Split(item, " (Device Group: ")
+			if len(parts) == 2 {
+				ruleName := parts[0]
+				dgPart := parts[1]
+
+				// Remove only the final closing parenthesis
+				dgPart = strings.TrimSuffix(dgPart, ")")
+
+				var deviceGroup, context string
+				if strings.Contains(dgPart, ", ") {
+					dgAndContext := strings.SplitN(dgPart, ", ", 2)
+					deviceGroup = dgAndContext[0]
+					context = dgAndContext[1]
+				} else {
+					deviceGroup = dgPart
+				}
+
+				rulesByDG[deviceGroup] = append(rulesByDG[deviceGroup], models.RuleContext{
+					Name:    ruleName,
+					Context: context,
+				})
+			}
+		}
+
+		// Sort device groups for consistent output
+		var deviceGroups []string
+		for dg := range rulesByDG {
+			deviceGroups = append(deviceGroups, dg)
+		}
+		sort.Strings(deviceGroups)
+
+		for _, dg := range deviceGroups {
+			rules := rulesByDG[dg]
+			fmt.Fprintf(file, "  %s:\n", dg)
+			for _, rule := range rules {
+				if rule.Context != "" {
+					fmt.Fprintf(file, "    - %s  # %s\n", rule.Name, rule.Context)
+				} else {
+					fmt.Fprintf(file, "    - %s\n", rule.Name)
+				}
+			}
+			fmt.Fprintf(file, "\n")
+		}
+	} else {
+		fmt.Fprintf(file, "  None discovered\n")
+	}
+	fmt.Fprintf(file, "\n")
+}
+
+func writeAddressGroupsCategory(file *os.File, groups []models.AddressGroup) {
+	count := len(groups)
+	fmt.Fprintf(file, "\n# ADDRESS GROUPS\n")
+	fmt.Fprintf(file, "# Found: %d item", count)
+	if count != 1 {
+		fmt.Fprintf(file, "s")
+	}
+	fmt.Fprintf(file, "\n---\n")
+
+	if count > 0 {
+		for i, group := range groups {
+			if group.Context == "shared" {
+				fmt.Fprintf(file, "  %d. %s (shared scope):\n", i+1, group.Name)
+				fmt.Fprintf(file, "     └─ Command: set shared address-group %s static %s\n", group.Name, group.Definition)
+				fmt.Fprintf(file, "     └─ Members: %s\n\n", group.Definition)
+			} else {
+				fmt.Fprintf(file, "  %d. %s (device-group - %s):\n", i+1, group.Name, group.DeviceGroup)
+				fmt.Fprintf(file, "     └─ Command: set device-group %s address-group %s static %s\n", group.DeviceGroup, group.Name, group.Definition)
+				fmt.Fprintf(file, "     └─ Members: %s\n\n", group.Definition)
+			}
+		}
+	} else {
+		fmt.Fprintf(file, "  None discovered\n")
+	}
+	fmt.Fprintf(file, "\n")
+}
+
+func writeRedundantAddressesCategory(file *os.File, addresses []models.RedundantAddress) {
+	count := len(addresses)
+	fmt.Fprintf(file, "\n# REDUNDANT ADDRESSES\n")
+	fmt.Fprintf(file, "# Found: %d item", count)
+	if count != 1 {
+		fmt.Fprintf(file, "s")
+	}
+	fmt.Fprintf(file, "\n---\n")
+
+	if count > 0 {
+		fmt.Fprintf(file, "  Address objects with identical IP configurations:\n\n")
+		for i, addr := range addresses {
+			fmt.Fprintf(file, "  %d. %s:\n", i+1, addr.Name)
+			fmt.Fprintf(file, "     └─ IP/Netmask: %s\n", addr.IPNetmask)
+			fmt.Fprintf(file, "     └─ Scope: %s\n", addr.DeviceGroup)
+			fmt.Fprintf(file, "     └─ Note: Same IP as target address - potential duplicate\n\n")
+		}
+	} else {
+		fmt.Fprintf(file, "  None discovered\n")
+	}
+	fmt.Fprintf(file, "\n")
+}
+
+// WriteAddressGroupCommands writes generated commands to a YAML file
+func WriteAddressGroupCommands(outputFile, originalAddress, newAddressName string, commands []string, addressGroups []models.AddressGroup) error {
+	// Ensure outputs directory exists
+	if err := EnsureOutputsDir(); err != nil {
+		return fmt.Errorf("error creating outputs directory: %w", err)
+	}
+	
+	// Prepend outputs/ to the file path if not already there
+	if !strings.HasPrefix(outputFile, "outputs/") {
+		outputFile = "outputs/" + outputFile
+	}
+	
+	file, err := os.Create(outputFile)
+	if err != nil {
+		return fmt.Errorf("error creating commands file: %w", err)
+	}
+	defer file.Close()
+
+	// Enhanced header with metadata
+	fmt.Fprintf(file, "# ═══════════════════════════════════════════════════════════════\n")
+	fmt.Fprintf(file, "# PAN Address Group Commands Generator v2.0 (Go Edition)\n")
+	fmt.Fprintf(file, "# ═══════════════════════════════════════════════════════════════\n")
+	fmt.Fprintf(file, "# Original Address Object: %s\n", originalAddress)
+	fmt.Fprintf(file, "# New Address Object: %s\n", newAddressName)
+	fmt.Fprintf(file, "# Address Groups Found: %d\n", len(addressGroups))
+	fmt.Fprintf(file, "# Commands Generated: %d\n", len(commands))
+	fmt.Fprintf(file, "# ═══════════════════════════════════════════════════════════════\n\n")
+
+	// Commands section
+	fmt.Fprintf(file, "# GENERATED CONFIGURATION COMMANDS\n")
+	fmt.Fprintf(file, "# Copy and paste these commands to add '%s' to the same address groups as '%s'\n", newAddressName, originalAddress)
+	fmt.Fprintf(file, "---\n\n")
+
+	if len(commands) > 0 {
+		for _, command := range commands {
+			fmt.Fprintf(file, "%s\n", command)
+		}
+	} else {
+		fmt.Fprintf(file, "# No commands generated\n")
+	}
+
+	// Address groups details section
+	fmt.Fprintf(file, "\n# SOURCE ADDRESS GROUPS\n")
+	fmt.Fprintf(file, "# These are the address groups that contained '%s'\n", originalAddress)
+	fmt.Fprintf(file, "---\n\n")
+
+	if len(addressGroups) > 0 {
+		for i, group := range addressGroups {
+			if group.Context == "shared" {
+				fmt.Fprintf(file, "  %d. %s (shared scope):\n", i+1, group.Name)
+				fmt.Fprintf(file, "     └─ Original Command: set shared address-group %s static %s\n", group.Name, group.Definition)
+				fmt.Fprintf(file, "     └─ New Command: set shared address-group %s static %s\n", group.Name, newAddressName)
+				fmt.Fprintf(file, "     └─ Members: %s\n\n", group.Definition)
+			} else {
+				fmt.Fprintf(file, "  %d. %s (device-group - %s):\n", i+1, group.Name, group.DeviceGroup)
+				fmt.Fprintf(file, "     └─ Original Command: set device-group %s address-group %s static %s\n", group.DeviceGroup, group.Name, group.Definition)
+				fmt.Fprintf(file, "     └─ New Command: set device-group %s address-group %s static %s\n", group.DeviceGroup, group.Name, newAddressName)
+				fmt.Fprintf(file, "     └─ Members: %s\n\n", group.Definition)
+			}
+		}
+	} else {
+		fmt.Fprintf(file, "  None discovered\n")
+	}
+
+	// Instructions section
+	fmt.Fprintf(file, "# USAGE INSTRUCTIONS\n")
+	fmt.Fprintf(file, "# 1. Create the new address object '%s' with appropriate IP/FQDN configuration\n", newAddressName)
+	fmt.Fprintf(file, "# 2. Copy the commands from the 'GENERATED CONFIGURATION COMMANDS' section above\n")
+	fmt.Fprintf(file, "# 3. Paste them into your PAN configuration interface or CLI\n")
+	fmt.Fprintf(file, "# 4. Commit the changes to apply the new address group memberships\n")
+	fmt.Fprintf(file, "---\n\n")
+
+	// Add footer
+	fmt.Fprintf(file, "# ═══════════════════════════════════════════════════════════════\n")
+	fmt.Fprintf(file, "# Commands Generation Complete\n")
+	fmt.Fprintf(file, "# Generated by: PAN Log Parser Tool v2.0 (Go Edition)\n")
+	fmt.Fprintf(file, "# Advanced Palo Alto Networks Configuration Analysis\n")
+	fmt.Fprintf(file, "# ═══════════════════════════════════════════════════════════════\n")
+
+	return nil
+}
